restructure graphql files and filename
