+++ server/package.json
@@ -12,6 +12,7 @@
 	"license": "ISC",
 	"dependencies": {
 		"@apollo/server": "^4.10.0",
+		"@graphql-tools/schema": "^10.0.2",
 		"bcrypt": "^5.1.1",
 		"express": "^4.18.2",
 		"graphql": "^16.8.1",
