+++ server/schemas/index.js
@@ -1,4 +1,12 @@
-const typeDefs = require('./typeDefs');
+const { mergeTypeDefs } = require('@graphql-tools/merge')
+
+const userTypeDefs = require('./userTypeDefs');
+const recommendationTypeDefs = require('./recommendationTypeDefs');
+
+const typeDefs = mergeTypeDefs([userTypeDefs, recommendationTypeDefs])
+
 const resolvers = require('./resolvers');
+
+
 
 module.exports = { typeDefs, resolvers };
